Clock ref 36

+++ BBBclockKernelPhysical/BBBclockKernelPhysicalDaemonSignalDriver.h
@@ -69,7 +69,7 @@ private:// Variables
 	static unsigned int *sharedMem_int,*pru0dataMem_int,*pru1dataMem_int;
 	// Time keeping
 	unsigned long long int TimeAdjPeriod=static_cast<unsigned long long int>(ClockCyclePeriodAdjustment*ClockPeriodNanoseconds); // Period at which the clock is adjusted. VEry important parameter
-	int TimePointClockCurrentAdjError=0;
+	double TimePointClockCurrentAdjError=0;
 	double TimePointClockCurrentAdjFilError=0;
 	double TimePointClockCurrentAdjFilErrorIntegral=0;
 	double TimePointClockCurrentAdjFilErrorDerivative=0;
